fix: logs copy button and progress bar display
- Fix copy/download logs fallback logic to use .strip() check
- Change current_epoch from int to float for accurate progress %
- Add tqdm progress bar parsing (100%|###| 33/33 format)
- Update epoch display to show fractional values
- Progress bar now correctly shows 100% when training completes

+++ ui/state.py
@@ -32,7 +32,7 @@ class JobState:
     completed_at: Optional[datetime] = None
     
     # Progress tracking
-    current_epoch: int = 0
+    current_epoch: float = 0.0  # Use float for fractional epochs (e.g., 2.75)
     total_epochs: int = 0
     current_step: int = 0
     total_steps: int = 0
